+++ vehicle/src/Vehicle/Compile.hs
@@ -8,6 +8,7 @@ module Vehicle.Compile
 where
 
 import Control.Monad.Writer (MonadWriter (..), WriterT (..))
+import Data.Set qualified as Set
 import Vehicle.Backend.ITP.Agda
 import Vehicle.Backend.ITP.Imandra
 import Vehicle.Backend.ITP.Isabelle
@@ -154,7 +155,7 @@ compileToLossFunction ::
   m ()
 compileToLossFunction LossOptions {..} typedProg outputAsJSON =
   logCompilerPass Loss $ do
-    lossTensorProg <- convertToLossTensors differentiableLogicID typedProg
+    lossTensorProg <- convertToLossTensors differentiableLogicID (Set.fromList declarationsToCompile) typedProg
     hoistedProg <- hoistInferableParameters lossTensorProg
     functionalisedProg <- functionaliseResources hoistedProg
     jsonProg <- convertToJSONProg functionalisedProg